+++ CHANGES.rst
@@ -8,6 +8,9 @@ Change Log
 
 * Required form fields are now labelled accordingly.
 
+* ``ralph_device_id`` get automatically cleaned when when Device linked to it gets deleted.
+
+
 1.2.6
 ~~~~~
 
@@ -20,6 +23,7 @@ Released on August 08, 2013
 * Added link to the Pricing App.
 
 * Added field: last modification, asset_id to csv file.
+
 
 1.0.0
 ~~~~~